Added memory storage implementation and a script for performance tests.

+++ src/main/java/net/troja/application/DataLoader.java
@@ -12,6 +12,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -23,15 +24,24 @@ public class DataLoader {
 
     @Autowired
     private Station2RouteRepository repository;
+    @Autowired
+    private ConnectionStorage storage;
     private int lines;
     private long entryCount;
+
+    @Value("${storage}")
+    private String storageType;
 
     public DataLoader() {
         super();
     }
 
     public void load(final Path file) {
-        repository.deleteAll();
+        if (BusRouteApplication.STORAGE_TYPE_DB.equals(storageType)) {
+            repository.deleteAll();
+        } else {
+            storage.clear();
+        }
         lines = 0;
 
         try (Stream<String> stream = Files.lines(file)) {
@@ -57,23 +67,45 @@ public class DataLoader {
         } catch (final IOException e) {
             LOGGER.error("Could not load data file", e);
         }
-        entryCount = repository.count();
+        if (BusRouteApplication.STORAGE_TYPE_DB.equals(storageType)) {
+            entryCount = repository.count();
+        } else {
+            entryCount = storage.size();
+        }
     }
 
     private void parseLine(final String line) {
         final String[] split = line.split(" ");
         try {
             if (split.length >= 3) {
-                final Integer route = Integer.parseInt(split[0]);
-                for (int pos = 1; pos < split.length; pos++) {
-                    final Integer station = Integer.parseInt(split[pos]);
-                    repository.save(new Station2Route(station, route));
+                if (BusRouteApplication.STORAGE_TYPE_DB.equals(storageType)) {
+                    parseLineToDb(split);
+                } else {
+                    parseLineToStorage(split);
                 }
             } else {
                 LOGGER.warn(LOG_INVALID_LINE);
             }
         } catch (final NumberFormatException e) {
             LOGGER.warn(LOG_INVALID_CONTENT);
+        }
+    }
+
+    private void parseLineToStorage(final String[] split) {
+        for (int pos = 1; pos < split.length; pos++) {
+            final Integer source = Integer.parseInt(split[pos]);
+            for (int run = pos; run < split.length; run++) {
+                final Integer destination = Integer.parseInt(split[run]);
+                storage.put(source, destination);
+            }
+        }
+    }
+
+    private void parseLineToDb(final String[] split) {
+        final Integer route = Integer.parseInt(split[0]);
+        for (int pos = 1; pos < split.length; pos++) {
+            final Integer station = Integer.parseInt(split[pos]);
+            repository.save(new Station2Route(station, route));
         }
     }
